requirements.txt
fastapifastapi==0.61.2
marshmallow==3.8.0
motor==2.3.0
pydantic==1.7.2
pymongo==3.11.1
ujson==4.0.1
urllib3==1.24.3
uvicorn==0.12.3
uvloop==0.14.0
